requirements.txt
google-genai
edge-tts
pygame
SpeechRecognition
pillow
pyautogui
python-dotenv